Use __class__.__name__ for boolops loopup
cmpops, binops, and unops already used this technique, but boolops did not.
The reason this technique is preferred to using __class__ for the lookup is that when working with e.g. gast (a cross version compatible AST) the exact type match required by the __class__ lookup fails (since gast.And != ast.And), whereas the name based lookup works fine.