Do a better job at handling very complex f-strings.
This is a change from the code copied from the Python code base, in
that, instead of naively repr'ing the "JoinedStr", we're checking the
string for quote types present.

There's gonna be a slow-down for very large strings, but I'm assuming
that it's not gonna be an issue in real-world code.